added admin controller and rerouted
