+++ app/helpers/admin/categories_helper.rb
@@ -1,0 +1,2 @@
+module Admin::CategoriesHelper
+end
