+++ app/views/admin/categories/index.html.erb
