+++ app/views/admin/categories/new.html.erb
@@ -1,0 +1,24 @@
+<%= provide(:title, "Add Category") %>
+<div class="container">
+    <div class="row">
+        <div class="col-lg-10 col-xl-9 mx-auto">
+            <div class="card card-signin flex-row my-5">
+                <div class="card-body">
+                    <h5 class="card-title text-center">New Category</h5>
+                    <%= form_for(@category, url: admin_category_path, html: {class: "form-signin"}) do |f|%>
+                        <div class="form-label-group">
+                            <%= f.text_field :title, class: "form-control", required: true, placeholder: "Title" %>
+                            <%= f.label :title %>
+                        </div>
+    
+                        <div class="form-label-group">
+                            <%= f.text_area :description, class: "form-control", required: true , id:"descriptionBox" ,placeholder:"Description", minlength: 50,maxlength: 150 %>
+                        </div>
+                        
+                        <%= f.submit "Add Category", class: "btn btn-lg btn-primary btn-block text-uppercase" %>    
+                    <% end %>  
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
