+++ test/controllers/admin/categories_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class Admin::CategoriesControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
